기본 API Sample 작성, Swagger-ui 적용

+++ src/main/java/com/skcc/tes/settlement/domain/SettlementRepository.java
@@ -2,6 +2,6 @@ package com.skcc.tes.settlement.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SettlementRepository extends JpaRepository<Settlement, Integer>{
+public interface SettlementRepository extends JpaRepository<Settlement, Long>{
     
 }
